@@ -1,6 +1,6 @@
 {
   "name": "mozaik-ext-time",
-  "version": "1.0.2",
+  "version": "2.0.0",
   "description": "Mozaik time widgets",
   "repository": {
     "type": "git",
@@ -12,29 +12,50 @@
     "url": "https://github.com/plouc"
   },
   "engines": {
-    "node": "0.12.x"
+    "node": ">=5.0.0",
+    "npm": ">=3.0.0"
   },
   "homepage": "https://github.com/plouc/mozaik-ext-time",
   "main": "./src/components/index.js",
+  "keywords": [
+    "time",
+    "mozaik",
+    "widget",
+    "extension",
+    "dashboard"
+  ],
   "dependencies": {
-    "babelify": "^6.1.3",
+    "classnames": "2.2.3",
     "d3": "^3.5.5",
+    "jquery": "^2.1.3",
     "moment": "^2.9.0",
     "moment-timezone": "^0.3.0",
-    "react-classset": "0.0.2",
-    "jquery": "^2.1.3",
-    "react": "^0.13.3",
-    "reflux": "^0.2.5"
+    "react": "0.14.7",
+    "react-mixin": "^1.7.0",
+    "reflux": "0.4.0"
   },
   "devDependencies": {
+    "babel-core": "6.7.2",
+    "babel-eslint": "5.0.0",
+    "babel-preset-es2015": "6.6.0",
+    "babel-preset-react": "6.5.0",
+    "babel-preset-stage-0": "6.5.0",
+    "babel-register": "6.7.2",
+    "enzyme": "2.1.0",
+    "eslint": "2.2.0",
+    "eslint-plugin-react": "4.2.2",
+    "expect": "1.15.2",
+    "mocha": "2.4.5",
+    "mockery": "1.4.1",
+    "react-addons-test-utils": "0.14.7",
+    "react-dom": "0.14.7",
+    "sinon": "1.17.3"
   },
   "peerDependencies": {
-    "mozaik": "1.x"
+    "mozaik": "1.2.x"
   },
-  "scripts": {},
-  "browserify": {
-    "transform": [
-      "babelify"
-    ]
+  "scripts": {
+    "eslint": "eslint --ext .js --ext .jsx ./src/*",
+    "test": "mocha --opts mocha.opts"
   }
 }
