@@ -1,8 +1,8 @@
-import React, { Component } from 'react';
-import classSet             from 'react-classset';
-import d3                   from 'd3/d3';
-import moment               from 'moment';
-import timezone             from 'moment-timezone';
+import React, { Component, PropTypes } from 'react';
+import classNames                      from 'classnames';
+import d3                              from 'd3/d3';
+import moment                          from 'moment';
+import timezone                        from 'moment-timezone';
 
 
 const getCurrentTimeParts = function (timezoneName) {
@@ -21,52 +21,63 @@ const secondsScale = d3.scale.linear().domain([0, 59 + 999/1000]).range([-90, 27
 const minutesScale = d3.scale.linear().domain([0, 59 + 59/60]).range([-90, 270]);
 const hoursScale   = d3.scale.linear().domain([0, 11 + 59/60]).range([-90, 270]);
 
+const sunFormats = ['HH:mm', 'H:mm', 'H:m'];
 
-var Clock = React.createClass({
-    getInitialState() {
-        return getCurrentTimeParts(this.props.timezone);
-    },
+
+class Clock extends Component {
+    static displayName = 'Clock';
+
+    static propTypes = {
+        title:    PropTypes.string,
+        info:     PropTypes.string,
+        timezone: PropTypes.string,
+        sunRise:  PropTypes.string.isRequired,
+        sunSet:   PropTypes.string.isRequired
+    };
+
+    static defaultProps = {
+        sunRise: '06:00',
+        sunSet:  '18:00'
+    };
+
+    constructor(props) {
+        super(props);
+
+        this.state = getCurrentTimeParts(this.props.timezone);
+    }
 
     componentDidMount() {
         setInterval(() => {
             this.setState(getCurrentTimeParts(this.props.timezone));
         }, 1000);
-    },
+    }
 
     render() {
-        var cs = classSet;
+        const { hours, minutes, seconds } = this.state;
+
         let hoursStyle   = {
-            transform: `rotate(${ hoursScale(this.state.hours % 12) }deg)`
+            transform: `rotate(${ hoursScale(hours % 12) }deg)`
         };
         let minutesStyle = {
-            transform: `rotate(${ minutesScale(this.state.minutes) }deg)`
+            transform: `rotate(${ minutesScale(minutes) }deg)`
         };
         let secondsStyle = {
-            transform: `rotate(${ secondsScale(this.state.seconds) }deg)`
+            transform: `rotate(${ secondsScale(seconds) }deg)`
         };
 
-        let { title } = this.props;
-
         // Day/night indicator
-        let sunRise = this.state.moment.clone().hours(6).minutes(0);
-        let sunSet  = this.state.moment.clone().hours(18).minutes(0);
+        const { sunRise, sunSet } = this.props;
 
-        // Parse custom times if given
-        if (this.props.sunRise) {
-            var customSunRiseTime = moment(this.props.sunRise, ['HH:mm', 'H:mm', 'H:m']);
-            sunRise.hours(customSunRiseTime.hours()).minutes(customSunRiseTime.minutes());
-        }
-        if (this.props.sunSet) {
-            var customSunSetTime = moment(this.props.sunSet, ['HH:mm', 'H:mm', 'H:m']);
-            sunSet.hours(customSunSetTime.hours()).minutes(customSunSetTime.minutes());
-        }
+        const sunRiseTime = moment(sunRise, sunFormats);
+        const sunSetTime  = moment(sunSet,  sunFormats);
 
-        const isDay = this.state.moment.isBetween(sunRise, sunSet);
-        const timeIndicatorClasses = cs({
-            'time__clock__indicator': true,
-            'fa':                     true,
-            'fa-sun-o':               isDay,
-            'fa-moon-o':              !isDay
+        const sunRiseDate = this.state.moment.clone().hours(sunRiseTime.hours()).minutes(sunRiseTime.minutes());
+        const sunSetDate  = this.state.moment.clone().hours(sunSetTime.hours()).minutes(sunSetTime.minutes());
+
+        const isDay = this.state.moment.isBetween(sunRiseDate, sunSetDate);
+        const timeIndicatorClasses = classNames('time__clock__indicator fa', {
+            'fa-sun-o':  isDay,
+            'fa-moon-o': !isDay
         });
 
         // Textual field, defaults to config value
@@ -77,12 +88,16 @@ var Clock = React.createClass({
         };
         const info = infoFields[this.props.info] || this.props.info;
 
+        let { title } = this.props;
+
         const bodyClasses = title ? 'widget__body' : '';
         const body = (
             <div className={bodyClasses}>
                 <div className="time__clock__outer-circle" />
                 <span className={timeIndicatorClasses}></span>
-                <span className="time__clock__info">{info}</span>
+                <span className="time__clock__info">
+                    {info}
+                </span>
                 <div className="time__clock__hand time__clock__hand--seconds" style={secondsStyle} />
                 <div className="time__clock__hand time__clock__hand--minutes" style={minutesStyle} />
                 <div className="time__clock__hand time__clock__hand--hours"   style={hoursStyle} />
@@ -108,6 +123,7 @@ var Clock = React.createClass({
             return body;
         }
     }
-});
+}
+
 
 export default Clock;
