@@ -2,5 +2,5 @@ import Clock from './Clock.jsx';
 
 
 export default {
-    Clock: Clock
+    Clock
 };
